Add all options present in other script (unimplemented)
including the options that may not be needed by MTM. Unneeded options
can still be removed later.

To do: make name, description and behavior of some options match with
other scripts, so that it is easier to understand each other's scripts
and to make it easy to see which options exist in others' scripts that
your script doesn't support.

This improves compatibility and encourages collaboration.